xege

Module mat

Modules

dynamic

Dynamic size matrix

Structs

Mat

A static size matrix of row primary order for an M x N.

Traits

IntoEGEMatrix

A trait for converting a other format matrix into an EGE matrix.

Functions

look_at

Generate a matrix of camera position transformations

ortho

Generate a orthographic projection matrix

perspective

Generate a perspective projection matrix

radian

Angle system to radian system

rotate2

Generate a 2D rotation transformation matrix.

rotate3

Generate a 3D rotation transformation matrix.

rotate3_x

A 3D rotation transformation matrix around the x-axis is generated

rotate3_y

A 3D rotation transformation matrix around the y-axis is generated

rotate3_z

A 3D rotation transformation matrix is generated about the z-axis

scale2

Generate a 2D scaled transformation matrix.

scale3

Generate a 3D scale transformation matrix

translate2

Generate a 2D translational transformation matrix.

translate3

Generate a 3D translational transformation matrix.

Type Aliases

Mat2

Mat3

Mat4

Mat2x3

Mat2x4

Mat3x2

Mat3x4

Mat4x2

Mat4x3

Vec

Vec2

Vec3

Vec4