xege::mat::dynamic

Struct Mat

pub struct Mat<T> { /* private fields */ }

A dynamic size matrix of row primary order for an M x N.

It has the same memory layout as Mat<T, M, N>, and can also interact with the API by taking the address of the element Mat[(0,0)].

Generic Parameters

T : The type of matrix element

Implementations

impl<T> Mat<T>

pub fn rows(&self) -> usize

Get the number of rows of the matrix.

pub fn cols(&self) -> usize

Get the number of columns for the matrix.

pub fn count(&self) -> usize

Get the number of elements of the matrix.

pub fn is_square(&self) -> bool

Determine if the matrix is a phalanx or not.

pub fn is_same_size_with(&self, other: &Mat<T>) -> bool

Determine whether the two matrices are the same size.

impl<T> Mat<T>

pub fn new(rows: usize, cols: usize) -> Mat<T>

impl<T> Mat<T>
where T: Copy,

pub fn new_with_init(rows: usize, cols: usize, init: T) -> Mat<T>

impl<T> Mat<T>
where T: PartialOrd + Neg<Output = T> + Sub<Output = T> + Copy,

pub fn eq_with_epsilon(&self, other: &Mat<T>, epsilon: T) -> bool

Equality judgment function for allowable errors.

impl<T> Mat<T>
where T: Copy,

pub fn transpose(&self) -> Mat<T>

Get a transpose matrix for the matrix.

pub fn T(&self) -> Mat<T>

It is an alias of transpose().

impl<T> Mat<T>
where T: Copy + From<f32>,

pub fn identity(size: usize) -> Mat<T>

Create an identity matrix of the specified size.

pub fn I(size: usize) -> Mat<T>

It is an alias of identity().

impl<T> Mat<T>
where T: Copy,

pub fn swap_row(&mut self, i: usize, j: usize) -> Result<(), &'static str>

Swap two rows of the matrix.

pub fn swap_col(&mut self, i: usize, j: usize) -> Result<(), &'static str>

Swap two columns of the matrix.

pub fn get_row(&self, i: usize) -> Result<Mat<T>, &'static str>

Get a copy of the elements of a row of the matrix.

pub fn get_col(&self, j: usize) -> Result<Mat<T>, &'static str>

Get a copy of the elements of a column of the matrix.

pub fn set_row(&mut self, i: usize, row: &Mat<T>) -> Result<(), &'static str>

Set the elements of a row of the matrix to specified vector.

pub fn set_col(&mut self, j: usize, col: &Mat<T>) -> Result<(), &'static str>

Set the elements of a column of the matrix to specified vector.

impl<T> Mat<T>
where T: PartialEq + Div<Output = T> + Mul<Output = T> + Copy + Sub<Output = T> + From<f32>,

pub fn lu(&self) -> Result<(Mat<T>, Mat<T>), &'static str>

LU decomposition is performed on the matrix.

Return

If it returns Ok(x): It will return a tuple containing two matrices: - The first matrix is the lower triangular matrix. - The second matrix is the upper triangular matrix.

pub fn det(&self) -> Result<T, &'static str>

Calculate the determinant value of the matrix.

Return

If it returns Ok(x): The determinant value of the matrix.

impl<T> Mat<T>
where T: Copy + From<f32> + PartialOrd + Debug + DivAssign + SubAssign + Mul<Output = T>,

pub fn inverse(&self) -> Result<Option<Mat<T>>, &'static str>

Calculate the inverse matrix of the matrix

Return

If it returns Ok(x): The function returns an Option:

If the matrix is invertible, it will return a Some(x) value containing the inverse matrix.

If the matrix is irreversible, it will return None.

impl<T> Mat<T>
where T: Default + PartialEq + SubAssign + Mul<Output = T> + Div<Output = T> + Copy,

pub fn rank(&self) -> Result<usize, &'static str>

Calculate the rank of the matrix

Return

If it returns Ok(x): The rank of the matrix