xege::mat

Struct Mat

pub struct Mat<T, const M: usize, const N: usize> { /* private fields */ }

A static size matrix of row primary order for an M x N.

Generic Parameters

T : The type of matrix element

M : The number of rows of the matrix

N : The number of columns in the matrix

Implementations

impl<T, const M: usize, const N: usize> Mat<T, M, N>
where T: Copy + PartialOrd + From<f32> + Sub<Output = T> + Neg<Output = T>,

pub fn eq_with_epsilon(&self, other: &Mat<T, M, N>, epsilon: T) -> bool

Matrix equality comparisons that allow floating-point errors.

impl<T, const M: usize, const N: usize> Mat<T, M, N>
where T: Copy,

pub fn new_with_init(init: T) -> Mat<T, M, N>

Create a matrix of M x N with the given value

Example

use mats::Mat;

let mat1 = Mat::<f32, 2,3>::new_with_init(1.0);
let mat2 = Mat::from([[1.0f32;3];2]);

assert_eq!(mat1, mat2);

impl<T, const M: usize, const N: usize> Mat<T, M, N>
where T: Default + Copy,

pub fn new() -> Mat<T, M, N>

Create a matrix of M x N with default values of T::default()

Example

use mats::Mat;

let mat1 = Mat::<f32, 2,3>::new();
let mat2 = Mat::from([[0.0f32;3];2]);

assert_eq!(mat1, mat2);

impl<T, const M: usize, const N: usize> Mat<T, M, N>
where T: Copy + Default,

pub fn T(&self) -> Mat<T, N, M>

Generates a transpose matrix for the current matrix

pub fn transpose(&self) -> Mat<T, N, M>

It is same as T() method.

impl<T, const M: usize> Mat<T, M, M>
where T: Copy + From<f32>,

pub fn I() -> Mat<T, M, M>

Generate an identity matrix

pub fn identity() -> Mat<T, M, M>

It is same as I() method.

impl<T, const M: usize, const N: usize> Mat<T, M, N>

pub const fn rows(&self) -> usize

Get the number of rows of the matrix.

pub const fn cols(&self) -> usize

Get the number of columns of the matrix.

pub const fn count(&self) -> usize

Get the number of elements of the matrix.

impl<T, const M: usize, const N: usize> Mat<T, M, N>
where T: Copy + Default,

pub fn swap_row(&mut self, i: usize, j: usize)

Swap two rows of the matrix.

Parameters

i,j : The row index which to be swapped.

pub fn swap_col(&mut self, i: usize, j: usize)

Swap two columns of the matrix.

Parameters

i,j : The column index which to be swapped.

impl<T, const M: usize> Mat<T, 1, M>
where T: Copy + Default + Into<f32> + AddAssign + Mul<Output = T>,

pub fn norm(&self) -> f32

Calculate the modulus length of the vector.

impl<T, const M: usize> Mat<T, 1, M>
where T: Copy + Default + Into<f32> + AddAssign + Mul<Output = T> + Div<Output = T> + From<f32>,

pub fn normalize(&self) -> Mat<T, 1, M>

Normalize vector.

impl<T> Mat<T, 1, 3>
where T: Default + Into<f32> + Mul<Output = T> + Sub<Output = T> + Copy,

pub fn cross(&self, other: Mat<T, 1, 3>) -> Mat<T, 1, 3>

Calculate the cross product of two vectors.

impl<T, const M: usize> Mat<T, M, M>
where T: From<f32> + PartialEq + Default + Mul<Output = T> + Sub<Output = T> + Div<Output = T> + Copy,

pub fn lu(&self) -> (Mat<T, M, M>, Mat<T, M, M>)

LU decomposition is performed on the matrix.

Return

It will return a tuple containing two matrices: - The first matrix is the lower triangular matrix. - The second matrix is the upper triangular matrix.

Example

use mats::Mat;

let a = Mat::from([
    [1.0, 2.0, 3.0],
    [4.0, 5.0, 6.0],
    [7.0, 8.0, 9.0]
]);
let (l, u) = a.lu();

assert_eq!(l, Mat::from([
    [1.0, 0.0, 0.0],
    [4.0, 1.0, 0.0],
    [7.0, 2.0, 1.0]
]));
assert_eq!(u, Mat::from([
    [1.0,  2.0,  3.0],
    [0.0, -3.0, -6.0],
    [0.0,  0.0,  0.0]
]));
assert_eq!(l * u, a);

pub fn det(&self) -> T

Calculate the determinant value of the matrix.

Return

The determinant value of the matrix.

Example

use mats::Mat;

let a = Mat::from([
    [1.0, 2.0, 3.0],
    [4.0, 5.0, 6.0],
    [7.0, 8.0, 9.0]
]);
assert_eq!(a.det(), 0.0);

impl<T, const M: usize> Mat<T, M, M>
where T: Copy + Default + From<f32> + PartialOrd + Debug + DivAssign + SubAssign + Mul<Output = T>,

pub fn inverse(&self) -> Option<Mat<T, M, M>>

Calculate the inverse matrix of the matrix

Return

The function returns an Option:

If the matrix is invertible, it will return a Some(x) value containing the inverse matrix.

If the matrix is irreversible, it will return None.

Example

use mats::Mat;

let a = Mat::from([
    [1.0, 2.0, 3.0],
    [4.0, 5.0, 6.0],
    [7.0, 8.0, 9.0]
]);
assert_eq!(a.inverse(), None);

let mat = Mat::from([
    [1.0, 2.0, 3.0],
    [4.0, 7.0, 6.0],
    [7.0, 8.0, 9.0]
]);
assert!(Mat::<f64,3,3>::I().eq_with_epsilon(&(mat * mat.inverse().unwrap()), 1e-6));

impl<T, const M: usize, const N: usize> Mat<T, M, N>
where T: Copy + Default + PartialEq + SubAssign + Mul<Output = T> + Div<Output = T>,

pub fn rank(&self) -> usize

Calculate the rank of the matrix

Return

The rank of the matrix

Example

use mats::{Mat3, Mat4};
let mat = Mat3::from([
    [1.0, 2.0, 3.0],
    [4.0, 5.0, 6.0],
    [7.0, 8.0, 9.0]
]);
assert_eq!(mat.rank(), 2);

assert_eq!(Mat4::<f32>::new().rank(), 0);

impl<T> Mat<T, 1, 2>

pub fn set_x(&mut self, value: T)

pub fn set_y(&mut self, value: T)

impl<T> Mat<T, 1, 3>

pub fn set_x(&mut self, value: T)

pub fn set_y(&mut self, value: T)

pub fn set_z(&mut self, value: T)

impl<T> Mat<T, 1, 4>

pub fn set_x(&mut self, value: T)

pub fn set_y(&mut self, value: T)

pub fn set_z(&mut self, value: T)

pub fn set_w(&mut self, value: T)

impl<T> Mat<T, 1, 2>
where T: Copy,

pub fn x(&self) -> T

GLSL syntax: v.x

pub fn y(&self) -> T

GLSL syntax: v.y

pub fn xx(&self) -> Mat<T, 1, 2>

GLSL syntax: v.xx

pub fn xy(&self) -> Mat<T, 1, 2>

GLSL syntax: v.xy

pub fn yx(&self) -> Mat<T, 1, 2>

GLSL syntax: v.yx

pub fn yy(&self) -> Mat<T, 1, 2>

GLSL syntax: v.yy

pub fn xxx(&self) -> Mat<T, 1, 3>

GLSL syntax: v.xxx

pub fn xxy(&self) -> Mat<T, 1, 3>

GLSL syntax: v.xxy

pub fn xyx(&self) -> Mat<T, 1, 3>

GLSL syntax: v.xyx

pub fn xyy(&self) -> Mat<T, 1, 3>

GLSL syntax: v.xyy

pub fn yxx(&self) -> Mat<T, 1, 3>

GLSL syntax: v.yxx

pub fn yxy(&self) -> Mat<T, 1, 3>

GLSL syntax: v.yxy

pub fn yyx(&self) -> Mat<T, 1, 3>

GLSL syntax: v.yyx

pub fn yyy(&self) -> Mat<T, 1, 3>

GLSL syntax: v.yyy

pub fn xxxx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xxxx

pub fn xxxy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xxxy

pub fn xxyx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xxyx

pub fn xxyy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xxyy

pub fn xyxx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xyxx

pub fn xyxy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xyxy

pub fn xyyx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xyyx

pub fn xyyy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xyyy

pub fn yxxx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yxxx

pub fn yxxy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yxxy

pub fn yxyx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yxyx

pub fn yxyy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yxyy

pub fn yyxx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yyxx

pub fn yyxy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yyxy

pub fn yyyx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yyyx

pub fn yyyy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yyyy

impl<T> Mat<T, 1, 3>
where T: Copy,

pub fn x(&self) -> T

GLSL syntax: v.x

pub fn y(&self) -> T

GLSL syntax: v.y

pub fn z(&self) -> T

GLSL syntax: v.z

pub fn xx(&self) -> Mat<T, 1, 2>

GLSL syntax: v.xx

pub fn xy(&self) -> Mat<T, 1, 2>

GLSL syntax: v.xy

pub fn xz(&self) -> Mat<T, 1, 2>

GLSL syntax: v.xz

pub fn yx(&self) -> Mat<T, 1, 2>

GLSL syntax: v.yx

pub fn yy(&self) -> Mat<T, 1, 2>

GLSL syntax: v.yy

pub fn yz(&self) -> Mat<T, 1, 2>

GLSL syntax: v.yz

pub fn zx(&self) -> Mat<T, 1, 2>

GLSL syntax: v.zx

pub fn zy(&self) -> Mat<T, 1, 2>

GLSL syntax: v.zy

pub fn zz(&self) -> Mat<T, 1, 2>

GLSL syntax: v.zz

pub fn xxx(&self) -> Mat<T, 1, 3>

GLSL syntax: v.xxx

pub fn xxy(&self) -> Mat<T, 1, 3>

GLSL syntax: v.xxy

pub fn xxz(&self) -> Mat<T, 1, 3>

GLSL syntax: v.xxz

pub fn xyx(&self) -> Mat<T, 1, 3>

GLSL syntax: v.xyx

pub fn xyy(&self) -> Mat<T, 1, 3>

GLSL syntax: v.xyy

pub fn xyz(&self) -> Mat<T, 1, 3>

GLSL syntax: v.xyz

pub fn xzx(&self) -> Mat<T, 1, 3>

GLSL syntax: v.xzx

pub fn xzy(&self) -> Mat<T, 1, 3>

GLSL syntax: v.xzy

pub fn xzz(&self) -> Mat<T, 1, 3>

GLSL syntax: v.xzz

pub fn yxx(&self) -> Mat<T, 1, 3>

GLSL syntax: v.yxx

pub fn yxy(&self) -> Mat<T, 1, 3>

GLSL syntax: v.yxy

pub fn yxz(&self) -> Mat<T, 1, 3>

GLSL syntax: v.yxz

pub fn yyx(&self) -> Mat<T, 1, 3>

GLSL syntax: v.yyx

pub fn yyy(&self) -> Mat<T, 1, 3>

GLSL syntax: v.yyy

pub fn yyz(&self) -> Mat<T, 1, 3>

GLSL syntax: v.yyz

pub fn yzx(&self) -> Mat<T, 1, 3>

GLSL syntax: v.yzx

pub fn yzy(&self) -> Mat<T, 1, 3>

GLSL syntax: v.yzy

pub fn yzz(&self) -> Mat<T, 1, 3>

GLSL syntax: v.yzz

pub fn zxx(&self) -> Mat<T, 1, 3>

GLSL syntax: v.zxx

pub fn zxy(&self) -> Mat<T, 1, 3>

GLSL syntax: v.zxy

pub fn zxz(&self) -> Mat<T, 1, 3>

GLSL syntax: v.zxz

pub fn zyx(&self) -> Mat<T, 1, 3>

GLSL syntax: v.zyx

pub fn zyy(&self) -> Mat<T, 1, 3>

GLSL syntax: v.zyy

pub fn zyz(&self) -> Mat<T, 1, 3>

GLSL syntax: v.zyz

pub fn zzx(&self) -> Mat<T, 1, 3>

GLSL syntax: v.zzx

pub fn zzy(&self) -> Mat<T, 1, 3>

GLSL syntax: v.zzy

pub fn zzz(&self) -> Mat<T, 1, 3>

GLSL syntax: v.zzz

pub fn xxxx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xxxx

pub fn xxxy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xxxy

pub fn xxxz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xxxz

pub fn xxyx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xxyx

pub fn xxyy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xxyy

pub fn xxyz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xxyz

pub fn xxzx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xxzx

pub fn xxzy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xxzy

pub fn xxzz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xxzz

pub fn xyxx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xyxx

pub fn xyxy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xyxy

pub fn xyxz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xyxz

pub fn xyyx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xyyx

pub fn xyyy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xyyy

pub fn xyyz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xyyz

pub fn xyzx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xyzx

pub fn xyzy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xyzy

pub fn xyzz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xyzz

pub fn xzxx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xzxx

pub fn xzxy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xzxy

pub fn xzxz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xzxz

pub fn xzyx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xzyx

pub fn xzyy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xzyy

pub fn xzyz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xzyz

pub fn xzzx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xzzx

pub fn xzzy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xzzy

pub fn xzzz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xzzz

pub fn yxxx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yxxx

pub fn yxxy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yxxy

pub fn yxxz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yxxz

pub fn yxyx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yxyx

pub fn yxyy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yxyy

pub fn yxyz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yxyz

pub fn yxzx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yxzx

pub fn yxzy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yxzy

pub fn yxzz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yxzz

pub fn yyxx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yyxx

pub fn yyxy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yyxy

pub fn yyxz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yyxz

pub fn yyyx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yyyx

pub fn yyyy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yyyy

pub fn yyyz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yyyz

pub fn yyzx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yyzx

pub fn yyzy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yyzy

pub fn yyzz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yyzz

pub fn yzxx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yzxx

pub fn yzxy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yzxy

pub fn yzxz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yzxz

pub fn yzyx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yzyx

pub fn yzyy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yzyy

pub fn yzyz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yzyz

pub fn yzzx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yzzx

pub fn yzzy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yzzy

pub fn yzzz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yzzz

pub fn zxxx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zxxx

pub fn zxxy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zxxy

pub fn zxxz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zxxz

pub fn zxyx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zxyx

pub fn zxyy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zxyy

pub fn zxyz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zxyz

pub fn zxzx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zxzx

pub fn zxzy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zxzy

pub fn zxzz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zxzz

pub fn zyxx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zyxx

pub fn zyxy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zyxy

pub fn zyxz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zyxz

pub fn zyyx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zyyx

pub fn zyyy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zyyy

pub fn zyyz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zyyz

pub fn zyzx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zyzx

pub fn zyzy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zyzy

pub fn zyzz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zyzz

pub fn zzxx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zzxx

pub fn zzxy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zzxy

pub fn zzxz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zzxz

pub fn zzyx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zzyx

pub fn zzyy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zzyy

pub fn zzyz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zzyz

pub fn zzzx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zzzx

pub fn zzzy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zzzy

pub fn zzzz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zzzz

impl<T> Mat<T, 1, 4>
where T: Copy,

pub fn x(&self) -> T

GLSL syntax: v.x

pub fn y(&self) -> T

GLSL syntax: v.y

pub fn z(&self) -> T

GLSL syntax: v.z

pub fn w(&self) -> T

GLSL syntax: v.w

pub fn xx(&self) -> Mat<T, 1, 2>

GLSL syntax: v.xx

pub fn xy(&self) -> Mat<T, 1, 2>

GLSL syntax: v.xy

pub fn xz(&self) -> Mat<T, 1, 2>

GLSL syntax: v.xz

pub fn xw(&self) -> Mat<T, 1, 2>

GLSL syntax: v.xw

pub fn yx(&self) -> Mat<T, 1, 2>

GLSL syntax: v.yx

pub fn yy(&self) -> Mat<T, 1, 2>

GLSL syntax: v.yy

pub fn yz(&self) -> Mat<T, 1, 2>

GLSL syntax: v.yz

pub fn yw(&self) -> Mat<T, 1, 2>

GLSL syntax: v.yw

pub fn zx(&self) -> Mat<T, 1, 2>

GLSL syntax: v.zx

pub fn zy(&self) -> Mat<T, 1, 2>

GLSL syntax: v.zy

pub fn zz(&self) -> Mat<T, 1, 2>

GLSL syntax: v.zz

pub fn zw(&self) -> Mat<T, 1, 2>

GLSL syntax: v.zw

pub fn wx(&self) -> Mat<T, 1, 2>

GLSL syntax: v.wx

pub fn wy(&self) -> Mat<T, 1, 2>

GLSL syntax: v.wy

pub fn wz(&self) -> Mat<T, 1, 2>

GLSL syntax: v.wz

pub fn ww(&self) -> Mat<T, 1, 2>

GLSL syntax: v.ww

pub fn xxx(&self) -> Mat<T, 1, 3>

GLSL syntax: v.xxx

pub fn xxy(&self) -> Mat<T, 1, 3>

GLSL syntax: v.xxy

pub fn xxz(&self) -> Mat<T, 1, 3>

GLSL syntax: v.xxz

pub fn xxw(&self) -> Mat<T, 1, 3>

GLSL syntax: v.xxw

pub fn xyx(&self) -> Mat<T, 1, 3>

GLSL syntax: v.xyx

pub fn xyy(&self) -> Mat<T, 1, 3>

GLSL syntax: v.xyy

pub fn xyz(&self) -> Mat<T, 1, 3>

GLSL syntax: v.xyz

pub fn xyw(&self) -> Mat<T, 1, 3>

GLSL syntax: v.xyw

pub fn xzx(&self) -> Mat<T, 1, 3>

GLSL syntax: v.xzx

pub fn xzy(&self) -> Mat<T, 1, 3>

GLSL syntax: v.xzy

pub fn xzz(&self) -> Mat<T, 1, 3>

GLSL syntax: v.xzz

pub fn xzw(&self) -> Mat<T, 1, 3>

GLSL syntax: v.xzw

pub fn xwx(&self) -> Mat<T, 1, 3>

GLSL syntax: v.xwx

pub fn xwy(&self) -> Mat<T, 1, 3>

GLSL syntax: v.xwy

pub fn xwz(&self) -> Mat<T, 1, 3>

GLSL syntax: v.xwz

pub fn xww(&self) -> Mat<T, 1, 3>

GLSL syntax: v.xww

pub fn yxx(&self) -> Mat<T, 1, 3>

GLSL syntax: v.yxx

pub fn yxy(&self) -> Mat<T, 1, 3>

GLSL syntax: v.yxy

pub fn yxz(&self) -> Mat<T, 1, 3>

GLSL syntax: v.yxz

pub fn yxw(&self) -> Mat<T, 1, 3>

GLSL syntax: v.yxw

pub fn yyx(&self) -> Mat<T, 1, 3>

GLSL syntax: v.yyx

pub fn yyy(&self) -> Mat<T, 1, 3>

GLSL syntax: v.yyy

pub fn yyz(&self) -> Mat<T, 1, 3>

GLSL syntax: v.yyz

pub fn yyw(&self) -> Mat<T, 1, 3>

GLSL syntax: v.yyw

pub fn yzx(&self) -> Mat<T, 1, 3>

GLSL syntax: v.yzx

pub fn yzy(&self) -> Mat<T, 1, 3>

GLSL syntax: v.yzy

pub fn yzz(&self) -> Mat<T, 1, 3>

GLSL syntax: v.yzz

pub fn yzw(&self) -> Mat<T, 1, 3>

GLSL syntax: v.yzw

pub fn ywx(&self) -> Mat<T, 1, 3>

GLSL syntax: v.ywx

pub fn ywy(&self) -> Mat<T, 1, 3>

GLSL syntax: v.ywy

pub fn ywz(&self) -> Mat<T, 1, 3>

GLSL syntax: v.ywz

pub fn yww(&self) -> Mat<T, 1, 3>

GLSL syntax: v.yww

pub fn zxx(&self) -> Mat<T, 1, 3>

GLSL syntax: v.zxx

pub fn zxy(&self) -> Mat<T, 1, 3>

GLSL syntax: v.zxy

pub fn zxz(&self) -> Mat<T, 1, 3>

GLSL syntax: v.zxz

pub fn zxw(&self) -> Mat<T, 1, 3>

GLSL syntax: v.zxw

pub fn zyx(&self) -> Mat<T, 1, 3>

GLSL syntax: v.zyx

pub fn zyy(&self) -> Mat<T, 1, 3>

GLSL syntax: v.zyy

pub fn zyz(&self) -> Mat<T, 1, 3>

GLSL syntax: v.zyz

pub fn zyw(&self) -> Mat<T, 1, 3>

GLSL syntax: v.zyw

pub fn zzx(&self) -> Mat<T, 1, 3>

GLSL syntax: v.zzx

pub fn zzy(&self) -> Mat<T, 1, 3>

GLSL syntax: v.zzy

pub fn zzz(&self) -> Mat<T, 1, 3>

GLSL syntax: v.zzz

pub fn zzw(&self) -> Mat<T, 1, 3>

GLSL syntax: v.zzw

pub fn zwx(&self) -> Mat<T, 1, 3>

GLSL syntax: v.zwx

pub fn zwy(&self) -> Mat<T, 1, 3>

GLSL syntax: v.zwy

pub fn zwz(&self) -> Mat<T, 1, 3>

GLSL syntax: v.zwz

pub fn zww(&self) -> Mat<T, 1, 3>

GLSL syntax: v.zww

pub fn wxx(&self) -> Mat<T, 1, 3>

GLSL syntax: v.wxx

pub fn wxy(&self) -> Mat<T, 1, 3>

GLSL syntax: v.wxy

pub fn wxz(&self) -> Mat<T, 1, 3>

GLSL syntax: v.wxz

pub fn wxw(&self) -> Mat<T, 1, 3>

GLSL syntax: v.wxw

pub fn wyx(&self) -> Mat<T, 1, 3>

GLSL syntax: v.wyx

pub fn wyy(&self) -> Mat<T, 1, 3>

GLSL syntax: v.wyy

pub fn wyz(&self) -> Mat<T, 1, 3>

GLSL syntax: v.wyz

pub fn wyw(&self) -> Mat<T, 1, 3>

GLSL syntax: v.wyw

pub fn wzx(&self) -> Mat<T, 1, 3>

GLSL syntax: v.wzx

pub fn wzy(&self) -> Mat<T, 1, 3>

GLSL syntax: v.wzy

pub fn wzz(&self) -> Mat<T, 1, 3>

GLSL syntax: v.wzz

pub fn wzw(&self) -> Mat<T, 1, 3>

GLSL syntax: v.wzw

pub fn wwx(&self) -> Mat<T, 1, 3>

GLSL syntax: v.wwx

pub fn wwy(&self) -> Mat<T, 1, 3>

GLSL syntax: v.wwy

pub fn wwz(&self) -> Mat<T, 1, 3>

GLSL syntax: v.wwz

pub fn www(&self) -> Mat<T, 1, 3>

GLSL syntax: v.www

pub fn xxxx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xxxx

pub fn xxxy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xxxy

pub fn xxxz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xxxz

pub fn xxxw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xxxw

pub fn xxyx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xxyx

pub fn xxyy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xxyy

pub fn xxyz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xxyz

pub fn xxyw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xxyw

pub fn xxzx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xxzx

pub fn xxzy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xxzy

pub fn xxzz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xxzz

pub fn xxzw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xxzw

pub fn xxwx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xxwx

pub fn xxwy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xxwy

pub fn xxwz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xxwz

pub fn xxww(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xxww

pub fn xyxx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xyxx

pub fn xyxy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xyxy

pub fn xyxz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xyxz

pub fn xyxw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xyxw

pub fn xyyx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xyyx

pub fn xyyy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xyyy

pub fn xyyz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xyyz

pub fn xyyw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xyyw

pub fn xyzx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xyzx

pub fn xyzy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xyzy

pub fn xyzz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xyzz

pub fn xyzw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xyzw

pub fn xywx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xywx

pub fn xywy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xywy

pub fn xywz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xywz

pub fn xyww(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xyww

pub fn xzxx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xzxx

pub fn xzxy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xzxy

pub fn xzxz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xzxz

pub fn xzxw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xzxw

pub fn xzyx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xzyx

pub fn xzyy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xzyy

pub fn xzyz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xzyz

pub fn xzyw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xzyw

pub fn xzzx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xzzx

pub fn xzzy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xzzy

pub fn xzzz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xzzz

pub fn xzzw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xzzw

pub fn xzwx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xzwx

pub fn xzwy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xzwy

pub fn xzwz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xzwz

pub fn xzww(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xzww

pub fn xwxx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xwxx

pub fn xwxy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xwxy

pub fn xwxz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xwxz

pub fn xwxw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xwxw

pub fn xwyx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xwyx

pub fn xwyy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xwyy

pub fn xwyz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xwyz

pub fn xwyw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xwyw

pub fn xwzx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xwzx

pub fn xwzy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xwzy

pub fn xwzz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xwzz

pub fn xwzw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xwzw

pub fn xwwx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xwwx

pub fn xwwy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xwwy

pub fn xwwz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xwwz

pub fn xwww(&self) -> Mat<T, 1, 4>

GLSL syntax: v.xwww

pub fn yxxx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yxxx

pub fn yxxy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yxxy

pub fn yxxz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yxxz

pub fn yxxw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yxxw

pub fn yxyx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yxyx

pub fn yxyy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yxyy

pub fn yxyz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yxyz

pub fn yxyw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yxyw

pub fn yxzx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yxzx

pub fn yxzy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yxzy

pub fn yxzz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yxzz

pub fn yxzw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yxzw

pub fn yxwx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yxwx

pub fn yxwy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yxwy

pub fn yxwz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yxwz

pub fn yxww(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yxww

pub fn yyxx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yyxx

pub fn yyxy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yyxy

pub fn yyxz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yyxz

pub fn yyxw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yyxw

pub fn yyyx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yyyx

pub fn yyyy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yyyy

pub fn yyyz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yyyz

pub fn yyyw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yyyw

pub fn yyzx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yyzx

pub fn yyzy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yyzy

pub fn yyzz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yyzz

pub fn yyzw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yyzw

pub fn yywx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yywx

pub fn yywy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yywy

pub fn yywz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yywz

pub fn yyww(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yyww

pub fn yzxx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yzxx

pub fn yzxy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yzxy

pub fn yzxz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yzxz

pub fn yzxw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yzxw

pub fn yzyx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yzyx

pub fn yzyy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yzyy

pub fn yzyz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yzyz

pub fn yzyw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yzyw

pub fn yzzx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yzzx

pub fn yzzy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yzzy

pub fn yzzz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yzzz

pub fn yzzw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yzzw

pub fn yzwx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yzwx

pub fn yzwy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yzwy

pub fn yzwz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yzwz

pub fn yzww(&self) -> Mat<T, 1, 4>

GLSL syntax: v.yzww

pub fn ywxx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.ywxx

pub fn ywxy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.ywxy

pub fn ywxz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.ywxz

pub fn ywxw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.ywxw

pub fn ywyx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.ywyx

pub fn ywyy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.ywyy

pub fn ywyz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.ywyz

pub fn ywyw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.ywyw

pub fn ywzx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.ywzx

pub fn ywzy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.ywzy

pub fn ywzz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.ywzz

pub fn ywzw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.ywzw

pub fn ywwx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.ywwx

pub fn ywwy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.ywwy

pub fn ywwz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.ywwz

pub fn ywww(&self) -> Mat<T, 1, 4>

GLSL syntax: v.ywww

pub fn zxxx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zxxx

pub fn zxxy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zxxy

pub fn zxxz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zxxz

pub fn zxxw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zxxw

pub fn zxyx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zxyx

pub fn zxyy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zxyy

pub fn zxyz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zxyz

pub fn zxyw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zxyw

pub fn zxzx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zxzx

pub fn zxzy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zxzy

pub fn zxzz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zxzz

pub fn zxzw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zxzw

pub fn zxwx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zxwx

pub fn zxwy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zxwy

pub fn zxwz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zxwz

pub fn zxww(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zxww

pub fn zyxx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zyxx

pub fn zyxy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zyxy

pub fn zyxz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zyxz

pub fn zyxw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zyxw

pub fn zyyx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zyyx

pub fn zyyy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zyyy

pub fn zyyz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zyyz

pub fn zyyw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zyyw

pub fn zyzx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zyzx

pub fn zyzy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zyzy

pub fn zyzz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zyzz

pub fn zyzw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zyzw

pub fn zywx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zywx

pub fn zywy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zywy

pub fn zywz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zywz

pub fn zyww(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zyww

pub fn zzxx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zzxx

pub fn zzxy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zzxy

pub fn zzxz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zzxz

pub fn zzxw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zzxw

pub fn zzyx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zzyx

pub fn zzyy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zzyy

pub fn zzyz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zzyz

pub fn zzyw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zzyw

pub fn zzzx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zzzx

pub fn zzzy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zzzy

pub fn zzzz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zzzz

pub fn zzzw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zzzw

pub fn zzwx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zzwx

pub fn zzwy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zzwy

pub fn zzwz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zzwz

pub fn zzww(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zzww

pub fn zwxx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zwxx

pub fn zwxy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zwxy

pub fn zwxz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zwxz

pub fn zwxw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zwxw

pub fn zwyx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zwyx

pub fn zwyy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zwyy

pub fn zwyz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zwyz

pub fn zwyw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zwyw

pub fn zwzx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zwzx

pub fn zwzy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zwzy

pub fn zwzz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zwzz

pub fn zwzw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zwzw

pub fn zwwx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zwwx

pub fn zwwy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zwwy

pub fn zwwz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zwwz

pub fn zwww(&self) -> Mat<T, 1, 4>

GLSL syntax: v.zwww

pub fn wxxx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wxxx

pub fn wxxy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wxxy

pub fn wxxz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wxxz

pub fn wxxw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wxxw

pub fn wxyx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wxyx

pub fn wxyy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wxyy

pub fn wxyz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wxyz

pub fn wxyw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wxyw

pub fn wxzx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wxzx

pub fn wxzy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wxzy

pub fn wxzz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wxzz

pub fn wxzw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wxzw

pub fn wxwx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wxwx

pub fn wxwy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wxwy

pub fn wxwz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wxwz

pub fn wxww(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wxww

pub fn wyxx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wyxx

pub fn wyxy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wyxy

pub fn wyxz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wyxz

pub fn wyxw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wyxw

pub fn wyyx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wyyx

pub fn wyyy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wyyy

pub fn wyyz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wyyz

pub fn wyyw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wyyw

pub fn wyzx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wyzx

pub fn wyzy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wyzy

pub fn wyzz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wyzz

pub fn wyzw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wyzw

pub fn wywx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wywx

pub fn wywy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wywy

pub fn wywz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wywz

pub fn wyww(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wyww

pub fn wzxx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wzxx

pub fn wzxy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wzxy

pub fn wzxz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wzxz

pub fn wzxw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wzxw

pub fn wzyx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wzyx

pub fn wzyy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wzyy

pub fn wzyz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wzyz

pub fn wzyw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wzyw

pub fn wzzx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wzzx

pub fn wzzy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wzzy

pub fn wzzz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wzzz

pub fn wzzw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wzzw

pub fn wzwx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wzwx

pub fn wzwy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wzwy

pub fn wzwz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wzwz

pub fn wzww(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wzww

pub fn wwxx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wwxx

pub fn wwxy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wwxy

pub fn wwxz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wwxz

pub fn wwxw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wwxw

pub fn wwyx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wwyx

pub fn wwyy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wwyy

pub fn wwyz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wwyz

pub fn wwyw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wwyw

pub fn wwzx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wwzx

pub fn wwzy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wwzy

pub fn wwzz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wwzz

pub fn wwzw(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wwzw

pub fn wwwx(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wwwx

pub fn wwwy(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wwwy

pub fn wwwz(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wwwz

pub fn wwww(&self) -> Mat<T, 1, 4>

GLSL syntax: v.wwww