xege

Trait Draw

pub trait Draw: DrawableDevice {

    // Provided methods
    fn setwritemode(&mut self, gen_rop2: impl Fn(u8, u8) -> u8) { ... }
    fn getpixel(&self, x: i32, y: i32) -> Color { ... }
    fn putpixel(&mut self, x: i32, y: i32, color: impl IntoARGB) { ... }
    fn putpixel_savealpha(&mut self, x: i32, y: i32, color: impl IntoARGB) { ... }
    fn putpixel_withalpha(&mut self, x: i32, y: i32, color: impl IntoARGB) { ... }
    fn putpixel_alphablend(
        &mut self,
        x: i32,
        y: i32,
        color: impl IntoARGB,
        factor: u8,
    ) { ... }
    fn putpixels<T: IntoARGB + Copy>(&mut self, array: &[PixelPoint<T>]) { ... }
    fn arc(&mut self, x: i32, y: i32, start: i32, end: i32, radius: i32) { ... }
    fn arcf(&mut self, x: f32, y: f32, start: f32, end: f32, radius: f32) { ... }
    fn ellipse(
        &mut self,
        x: i32,
        y: i32,
        start: i32,
        end: i32,
        rx: i32,
        ry: i32,
    ) { ... }
    fn ellipsef(
        &mut self,
        x: f32,
        y: f32,
        start: f32,
        end: f32,
        rx: f32,
        ry: f32,
    ) { ... }
    fn line(&mut self, x1: i32, y1: i32, x2: i32, y2: i32) { ... }
    fn linerel(&mut self, dx: i32, dy: i32) { ... }
    fn lineto(&mut self, x: i32, y: i32) { ... }
    fn drawlines(&mut self, lines: &[Line]) { ... }
    fn polyline(&mut self, points: &[Point]) { ... }
    fn drawpoly(&mut self, points: &[Point]) { ... }
    fn drawbezier(&mut self, points: &[Point]) { ... }
    fn rectangle(&mut self, left: i32, top: i32, right: i32, bottom: i32) { ... }
    fn fillrect(&mut self, left: i32, top: i32, right: i32, bottom: i32) { ... }
    fn solidrectangle(&mut self, left: i32, top: i32, right: i32, bottom: i32) { ... }
    fn bar(&mut self, left: i32, top: i32, right: i32, bottom: i32) { ... }
    fn bar3d(
        &mut self,
        left: i32,
        top: i32,
        right: i32,
        bottom: i32,
        depth: i32,
        topflag: bool,
    ) { ... }
    fn polygon(&mut self, points: &[Point]) { ... }
    fn fillpoly(&mut self, points: &[Point]) { ... }
    fn solidpoly(&mut self, points: &[Point]) { ... }
    fn fillpoly_gradient<T: IntoARGB + Copy>(
        &mut self,
        points: &[PixelPoint<T, f32>],
    ) { ... }
    fn circle(&mut self, x: i32, y: i32, radius: i32) { ... }
    fn circlef(&mut self, x: f32, y: f32, radius: f32) { ... }
    fn fillcircle(&mut self, x: i32, y: i32, radius: i32) { ... }
    fn fillcirclef(&mut self, x: f32, y: f32, radius: f32) { ... }
    fn fillellipse(&mut self, x: i32, y: i32, rx: i32, ry: i32) { ... }
    fn fillellipsef(&mut self, x: f32, y: f32, rx: f32, ry: f32) { ... }
    fn solidcircle(&mut self, x: i32, y: i32, radius: i32) { ... }
    fn solidcirclef(&mut self, x: f32, y: f32, radius: f32) { ... }
    fn solidellipse(&mut self, x: i32, y: i32, rx: i32, ry: i32) { ... }
    fn solidellipsef(&mut self, x: f32, y: f32, rx: f32, ry: f32) { ... }
    fn pie(&mut self, x: i32, y: i32, start: i32, end: i32, rx: i32, ry: i32) { ... }
    fn pief(&mut self, x: f32, y: f32, start: f32, end: f32, rx: f32, ry: f32) { ... }
    fn fillpie(
        &mut self,
        x: i32,
        y: i32,
        start: i32,
        end: i32,
        rx: i32,
        ry: i32,
    ) { ... }
    fn fillpief(
        &mut self,
        x: f32,
        y: f32,
        start: f32,
        end: f32,
        rx: f32,
        ry: f32,
    ) { ... }
    fn solidpie(
        &mut self,
        x: i32,
        y: i32,
        start: i32,
        end: i32,
        rx: i32,
        ry: i32,
    ) { ... }
    fn solidpief(
        &mut self,
        x: f32,
        y: f32,
        start: f32,
        end: f32,
        rx: f32,
        ry: f32,
    ) { ... }
    fn sector(&mut self, x: i32, y: i32, start: i32, end: i32, rx: i32, ry: i32) { ... }
    fn sectorf(
        &mut self,
        x: f32,
        y: f32,
        start: f32,
        end: f32,
        rx: f32,
        ry: f32,
    ) { ... }
    fn pieslice(&mut self, x: i32, y: i32, start: i32, end: i32, radius: i32) { ... }
    fn pieslicef(&mut self, x: f32, y: f32, start: f32, end: f32, radius: f32) { ... }
    fn roundrect(
        &mut self,
        left: i32,
        top: i32,
        right: i32,
        bottom: i32,
        rx: i32,
        ry: i32,
    ) { ... }
    fn fillroundrect(
        &mut self,
        left: i32,
        top: i32,
        right: i32,
        bottom: i32,
        rx: i32,
        ry: i32,
    ) { ... }
    fn solidroundrect(
        &mut self,
        left: i32,
        top: i32,
        right: i32,
        bottom: i32,
        rx: i32,
        ry: i32,
    ) { ... }
    fn floodfill(&mut self, x: i32, y: i32, color: impl IntoARGB) { ... }
    fn floodfillsurface(&mut self, x: i32, y: i32, color: impl IntoARGB) { ... }
    fn outtext(&mut self, text: &str) { ... }
    fn outtextrect(&mut self, x: i32, y: i32, w: i32, h: i32, text: &str) { ... }
    fn outtextxy(&mut self, x: i32, y: i32, text: &str) { ... }

}

Ordinary drawing functions.

Note

They do not support anti aliasing and transparent channels.

Provided Methods

fn setwritemode(&mut self, gen_rop2: impl Fn(u8, u8) -> u8)

Set ROP2.

Parameters

gen_rop2: The ROP2 generator.

Note

The gen_rop2 accepts two u8: - Pen: current pen color. - Dst: Screen color.

This function only valid for trait Draw.

Example

use xege::*;

let mut xege = initgraph(640, 480, Init::Default).unwarp();
// DPan
xege.setwritemode(|pen, dst| pen & dst);

fn getpixel(&self, x: i32, y: i32) -> Color

Get the color of the pixel at the specified position.

Parameters

x - The x position.

y - The y position.

Return

The color of the pixel.

fn putpixel(&mut self, x: i32, y: i32, color: impl IntoARGB)

Draw a pixel at the specified position.

Parameters

x - The x position.

y - The y position.

color - The color to draw.

fn putpixel_savealpha(&mut self, x: i32, y: i32, color: impl IntoARGB)

Set the color of pixels while preserving the original alpha value

Parameters

x - The x position.

y - The y position.

color - The color to draw.

fn putpixel_withalpha(&mut self, x: i32, y: i32, color: impl IntoARGB)

Draw pixel points Mix according to alpha, and retain the background color alpha in the mixed result.

Parameters

x - The x position.

y - The y position.

color - The color to draw.

fn putpixel_alphablend( &mut self, x: i32, y: i32, color: impl IntoARGB, factor: u8, )

Draw pixel points Mix based on alpha

Parameters

x - The x position.

y - The y position.

color - The color to draw.

factor - The alpha factor.

fn putpixels<T: IntoARGB + Copy>(&mut self, array: &[PixelPoint<T>])

Draw multiple pixels

Parameters

array - The array of pixels to draw.

fn arc(&mut self, x: i32, y: i32, start: i32, end: i32, radius: i32)

Draw a arc.

Parameters

x - The x position of the center.

y - The y position of the center.

start - The start angle in degrees.

end - The end angle in degrees.

radius - The radius of the arc.

fn arcf(&mut self, x: f32, y: f32, start: f32, end: f32, radius: f32)

Draw a arc.

Parameters

x - The x position of the center.

y - The y position of the center.

start - The start angle in degrees.

end - The end angle in degrees.

radius - The radius of the arc.

fn ellipse(&mut self, x: i32, y: i32, start: i32, end: i32, rx: i32, ry: i32)

Draw a ellipse.

Parameters

x - The x position of the center.

y - The y position of the center.

start - The start angle in degrees.

end - The end angle in degrees.

rx - The x radius of the ellipse.

ry - The y radius of the ellipse.

fn ellipsef(&mut self, x: f32, y: f32, start: f32, end: f32, rx: f32, ry: f32)

Draw a ellipse.

Parameters

x - The x position of the center.

y - The y position of the center.

start - The start angle in degrees.

end - The end angle in degrees.

rx - The x radius of the ellipse.

ry - The y radius of the ellipse.

fn line(&mut self, x1: i32, y1: i32, x2: i32, y2: i32)

Draw a line between two points.

Parameters

x1 - The x position of the first point.

y1 - The y position of the first point.

x2 - The x position of the second point.

y2 - The y position of the second point.

fn linerel(&mut self, dx: i32, dy: i32)

Draw a line between current position and the specified relative coordinates.

Parameters

dx - The relative x position.

dy - The relative y position.

fn lineto(&mut self, x: i32, y: i32)

Draw a line between current position and the specified absolute coordinates.

Parameters

x - The absolute x position.

y - The absolute y position.

fn drawlines(&mut self, lines: &[Line])

Draw multiple lines.

Parameters

lines - The array of lines to draw.

fn polyline(&mut self, points: &[Point])

Draw a polyline.

Parameters

points - The array of points to draw.

fn drawpoly(&mut self, points: &[Point])

Draw a polyline.

Parameters

points - The array of points to draw.

fn drawbezier(&mut self, points: &[Point])

Draw a bezier curve.

Parameters

points - The array of control points to draw.

fn rectangle(&mut self, left: i32, top: i32, right: i32, bottom: i32)

Draw a rectangle.

Parameters

left - The left position of the rectangle.

top - The top position of the rectangle.

right - The right position of the rectangle.

bottom - The bottom position of the rectangle.

fn fillrect(&mut self, left: i32, top: i32, right: i32, bottom: i32)

Draw a fill rectangle.

Parameters

left - The left position of the rectangle.

top - The top position of the rectangle.

right - The right position of the rectangle.

bottom - The bottom position of the rectangle.

fn solidrectangle(&mut self, left: i32, top: i32, right: i32, bottom: i32)

Draw a solid rectangle.

Parameters

left - The left position of the rectangle.

top - The top position of the rectangle.

right - The right position of the rectangle.

bottom - The bottom position of the rectangle.

fn bar(&mut self, left: i32, top: i32, right: i32, bottom: i32)

Draw a rectangle.

Parameters

left - The left position of the rectangle.

top - The top position of the rectangle.

right - The right position of the rectangle.

bottom - The bottom position of the rectangle.

fn bar3d( &mut self, left: i32, top: i32, right: i32, bottom: i32, depth: i32, topflag: bool, )

Draw a block.

Parameters

left - The left position of the block.

top - The top position of the block.

right - The right position of the block.

bottom - The bottom position of the block.

depth - The depth of the block.

topflag - Whether to fill the top of the block.

fn polygon(&mut self, points: &[Point])

Draw a polygon.

Parameters

points - The vertices of the polygon.

fn fillpoly(&mut self, points: &[Point])

Draw a filled polygon.

Parameters

points - The vertices of the polygon.

fn solidpoly(&mut self, points: &[Point])

Draw a solid polygon.

Parameters

points - The vertices of the polygon.

fn fillpoly_gradient<T: IntoARGB + Copy>( &mut self, points: &[PixelPoint<T, f32>], )

Draw a gradient polygon.

Parameters

points - The vertices of the polygon.

fn circle(&mut self, x: i32, y: i32, radius: i32)

Draw a circle.

Parameters

x - The x position of the center.

y - The y position of the center.

radius - The radius of the circle.

fn circlef(&mut self, x: f32, y: f32, radius: f32)

Draw a circle.

Parameters

x - The x position of the center.

y - The y position of the center.

radius - The radius of the circle.

fn fillcircle(&mut self, x: i32, y: i32, radius: i32)

Draw a filled circle.

Parameters

x - The x position of the center.

y - The y position of the center.

radius - The radius of the circle.

fn fillcirclef(&mut self, x: f32, y: f32, radius: f32)

Draw a filled circle.

Parameters

x - The x position of the center.

y - The y position of the center.

radius - The radius of the circle.

fn fillellipse(&mut self, x: i32, y: i32, rx: i32, ry: i32)

Draw a ellipse.

Parameters

x - The x position of the center.

y - The y position of the center.

rx - The x radius of the ellipse.

ry - The y radius of the ellipse.

fn fillellipsef(&mut self, x: f32, y: f32, rx: f32, ry: f32)

Draw a ellipse.

Parameters

x - The x position of the center.

y - The y position of the center.

rx - The x radius of the ellipse.

ry - The y radius of the ellipse.

fn solidcircle(&mut self, x: i32, y: i32, radius: i32)

Draw a solid circle.

Parameters

x - The x position of the center.

y - The y position of the center.

radius - The radius of the circle.

fn solidcirclef(&mut self, x: f32, y: f32, radius: f32)

Draw a solid circle.

Parameters

x - The x position of the center.

y - The y position of the center.

radius - The radius of the circle.

fn solidellipse(&mut self, x: i32, y: i32, rx: i32, ry: i32)

Draw a solid ellipse.

Parameters

x - The x position of the center.

y - The y position of the center.

rx - The x radius of the ellipse.

ry - The y radius of the ellipse.

fn solidellipsef(&mut self, x: f32, y: f32, rx: f32, ry: f32)

Draw a solid ellipse.

Parameters

x - The x position of the center.

y - The y position of the center.

rx - The x radius of the ellipse.

ry - The y radius of the ellipse.

fn pie(&mut self, x: i32, y: i32, start: i32, end: i32, rx: i32, ry: i32)

Draw a pie.

Parameters

x - The x position of the center.

y - The y position of the center.

start - The start angle of the pie.

end - The end angle of the pie.

rx - The x radius of the pie.

ry - The y radius of the pie.

fn pief(&mut self, x: f32, y: f32, start: f32, end: f32, rx: f32, ry: f32)

Draw a pie.

Parameters

x - The x position of the center.

y - The y position of the center.

start - The start angle of the pie.

end - The end angle of the pie.

rx - The x radius of the pie.

ry - The y radius of the pie.

fn fillpie(&mut self, x: i32, y: i32, start: i32, end: i32, rx: i32, ry: i32)

Draw a filled pie.

Parameters

x - The x position of the center.

y - The y position of the center.

start - The start angle of the pie.

end - The end angle of the pie.

rx - The x radius of the pie.

ry - The y radius of the pie.

fn fillpief(&mut self, x: f32, y: f32, start: f32, end: f32, rx: f32, ry: f32)

Draw a filled pie.

Parameters

x - The x position of the center.

y - The y position of the center.

start - The start angle of the pie.

end - The end angle of the pie.

rx - The x radius of the pie.

ry - The y radius of the pie.

fn solidpie(&mut self, x: i32, y: i32, start: i32, end: i32, rx: i32, ry: i32)

Draw a solid pie.

Parameters

x - The x position of the center.

y - The y position of the center.

start - The start angle of the pie.

end - The end angle of the pie.

rx - The x radius of the pie.

ry - The y radius of the pie.

fn solidpief(&mut self, x: f32, y: f32, start: f32, end: f32, rx: f32, ry: f32)

Draw a solid pie.

Parameters

x - The x position of the center.

y - The y position of the center.

start - The start angle of the pie.

end - The end angle of the pie.

rx - The x radius of the pie.

ry - The y radius of the pie.

fn sector(&mut self, x: i32, y: i32, start: i32, end: i32, rx: i32, ry: i32)

Draw a sector.

Parameters

x - The x position of the center.

y - The y position of the center.

start - The start angle of the sector.

end - The end angle of the sector.

rx - The x radius of the sector.

ry - The y radius of the sector.

fn sectorf(&mut self, x: f32, y: f32, start: f32, end: f32, rx: f32, ry: f32)

Draw a sector.

Parameters

x - The x position of the center.

y - The y position of the center.

start - The start angle of the sector.

end - The end angle of the sector.

rx - The x radius of the sector.

ry - The y radius of the sector.

fn pieslice(&mut self, x: i32, y: i32, start: i32, end: i32, radius: i32)

Draw a pieslice.

Parameters

x - The x position of the center.

y - The y position of the center.

start - The start angle of the pieslice.

end - The end angle of the pieslice.

radius - The radius of the pieslice.

fn pieslicef(&mut self, x: f32, y: f32, start: f32, end: f32, radius: f32)

Draw a pieslice.

Parameters

x - The x position of the center.

y - The y position of the center.

start - The start angle of the pieslice.

end - The end angle of the pieslice.

radius - The radius of the pieslice.

fn roundrect( &mut self, left: i32, top: i32, right: i32, bottom: i32, rx: i32, ry: i32, )

Draw a rounded rectangle.

Parameters

left - The left position of the rectangle.

top - The top position of the rectangle.

right - The right position of the rectangle.

bottom - The bottom position of the rectangle.

rx - The x radius of the rounded corners.

ry - The y radius of the rounded corners.

fn fillroundrect( &mut self, left: i32, top: i32, right: i32, bottom: i32, rx: i32, ry: i32, )

Draw a fill rounded rectangle.

Parameters

left - The left position of the rectangle.

top - The top position of the rectangle.

right - The right position of the rectangle.

bottom - The bottom position of the rectangle.

rx - The x radius of the rounded corners.

ry - The y radius of the rounded corners.

fn solidroundrect( &mut self, left: i32, top: i32, right: i32, bottom: i32, rx: i32, ry: i32, )

Draw a solid rounded rectangle.

Parameters

left - The left position of the rectangle.

top - The top position of the rectangle.

right - The right position of the rectangle.

bottom - The bottom position of the rectangle.

rx - The x radius of the rounded corners.

ry - The y radius of the rounded corners.

fn floodfill(&mut self, x: i32, y: i32, color: impl IntoARGB)

Specify the boundary color filling area.

Parameters

x - Fill the x-position of any point within the region.

y - Fill the y-position of any point within the region.

color - Fill the color of the boundary of the area.

fn floodfillsurface(&mut self, x: i32, y: i32, color: impl IntoARGB)

Specify the surface color filling area.

Parameters

x - Fill the x-position of any point within the region.

y - Fill the y-position of any point within the region.

color - Fill the color of the surface of the area.

fn outtext(&mut self, text: &str)

Output text at current position.

Parameters

text - The text to output.

fn outtextrect(&mut self, x: i32, y: i32, w: i32, h: i32, text: &str)

Output text at specified rectangle area.

Parameters

x - The x position of the top-left corner.

y - The y position of the top-left corner.

w - The width of the rectangle.

h - The height of the rectangle.

text - The text to output.

fn outtextxy(&mut self, x: i32, y: i32, text: &str)

Output text at specified position (x, y).

Parameters

x - The x position.

y - The y position.

text - The text to output.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.