xege

Trait ImageDraw

pub trait ImageDraw: DrawableDevice {

    // Provided methods
    fn putimage(
        &mut self,
        x: i32,
        y: i32,
        image: &Image,
        gen_rop3: impl Fn(u32, u32, u32) -> u32,
    ) { ... }
    fn putimage_with_size(
        &mut self,
        dest: Rect,
        image: &Image,
        x_src: i32,
        y_src: i32,
        gen_rop3: impl Fn(u32, u32, u32) -> u32,
    ) { ... }
    fn putimage_with_scale(
        &mut self,
        dest: Rect,
        image: &Image,
        src: Rect,
        gen_rop3: impl Fn(u32, u32, u32) -> u32,
    ) { ... }
    fn putimage_with_alpha(
        &mut self,
        image: &Image,
        dest: Rect,
        src: Rect,
        smooth: bool,
    ) -> Result<(), ImageError> { ... }
    fn putimage_alphablender(
        &mut self,
        image: &Image,
        dest: Rect,
        alpha: u8,
        src: Rect,
        smooth: bool,
        alpha_type: AlphaType,
    ) -> Result<(), ImageError> { ... }
    fn putimage_alphafilter(
        &mut self,
        image: &Image,
        x_dest: i32,
        y_dest: i32,
        alpha: &Image,
        src: Rect,
    ) -> Result<(), ImageError> { ... }
    fn putimage_transparent(
        &mut self,
        image: &Image,
        x_dest: i32,
        y_dest: i32,
        transparent: impl IntoARGB,
        src: Rect,
    ) -> Result<(), ImageError> { ... }
    fn putimage_alphatransparent(
        &mut self,
        image: &Image,
        x_dest: i32,
        y_dest: i32,
        transparent: impl IntoARGB,
        alpha: u8,
        src: Rect,
    ) -> Result<(), ImageError> { ... }
    fn putimage_rotate(
        &mut self,
        image: &Image,
        x_dest: i32,
        y_dest: i32,
        x_center: f32,
        y_center: f32,
        radian: f32,
        use_alpha: bool,
        alpha: Option<u8>,
        smooth: bool,
    ) -> Result<(), ImageError> { ... }
    fn putimage_rotatezoom(
        &mut self,
        image: &Image,
        x_dest: i32,
        y_dest: i32,
        x_center: f32,
        y_center: f32,
        radian: f32,
        zoom: f32,
        use_alpha: bool,
        alpha: Option<u8>,
        smooth: bool,
    ) -> Result<(), ImageError> { ... }
    fn putimage_rotatetransparent(
        &mut self,
        image: &Image,
        x_center_dest: i32,
        y_center_dest: i32,
        src: Rect,
        x_center_src: i32,
        y_center_src: i32,
        transparent: impl IntoARGB,
        radian: f32,
        zoom: f32,
    ) -> Result<(), ImageError> { ... }
    fn imagefilter_blurring(
        &mut self,
        intensity: i32,
        alpha: u32,
        dest: Option<Rect>,
    ) -> Result<(), ImageError> { ... }
    fn drawimage(&mut self, image: &Image, x: i32, y: i32) { ... }
    fn drawimage_with_scale(&mut self, image: &Image, dest: Rect, src: Rect) { ... }

}

Provided Methods

fn putimage( &mut self, x: i32, y: i32, image: &Image, gen_rop3: impl Fn(u32, u32, u32) -> u32, )

Draw an image.

Parameters

image - The image to draw.

x - The x position of the top-left corner.

y - The y position of the top-left corner.

gen_rop3 - A function to generate the ROP3 code.

Note

The gen_rop3 function takes three parameters: - Pen mask. - Src mask. - Dst mask.

For example, the DPo is |P, S, D| D | P, the DPSao is |P, S, D| (S & P) | D, and so on.

fn putimage_with_size( &mut self, dest: Rect, image: &Image, x_src: i32, y_src: i32, gen_rop3: impl Fn(u32, u32, u32) -> u32, )

Draw an image with a specified size.

Parameters

dest - The destination rectangle.

image - The image to draw.

x_src - The x position of the top-left corner of the source image.

y_src - The y position of the top-left corner of the source image.

gen_rop3 - A function to generate the ROP3 code.

fn putimage_with_scale( &mut self, dest: Rect, image: &Image, src: Rect, gen_rop3: impl Fn(u32, u32, u32) -> u32, )

Draw an image with a specified scale.

Parameters

dest - The destination rectangle.

image - The image to draw.

src - The source rectangle.

gen_rop3 - A function to generate the ROP3 code.

fn putimage_with_alpha( &mut self, image: &Image, dest: Rect, src: Rect, smooth: bool, ) -> Result<(), ImageError>

Draw an image with alpha.

Parameters

image - The image to draw.

dest - The destination rectangle.

src - The source rectangle.

smooth - Whether to use smooth scaling.

fn putimage_alphablender( &mut self, image: &Image, dest: Rect, alpha: u8, src: Rect, smooth: bool, alpha_type: AlphaType, ) -> Result<(), ImageError>

Draw an image with alpha blending and a specified alpha type.

Parameters

image - The image to draw.

dest - The destination rectangle.

alpha - The alpha value.

src - The source rectangle.

smooth - Whether to use smooth scaling.

alpha_type - The alpha type.

fn putimage_alphafilter( &mut self, image: &Image, x_dest: i32, y_dest: i32, alpha: &Image, src: Rect, ) -> Result<(), ImageError>

Draw an image with alpha blending and a specified alpha type.

Parameters

image - The image to draw.

x_dest - The x position of the top-left corner of the destination image.

y_dest - The y position of the top-left corner of the destination image.

alpha - The alpha image.

src - The source rectangle.

fn putimage_transparent( &mut self, image: &Image, x_dest: i32, y_dest: i32, transparent: impl IntoARGB, src: Rect, ) -> Result<(), ImageError>

Draw an image with alpha blending and a specified alpha type.

Parameters

image - The image to draw.

x_dest - The x position of the top-left corner of the destination image.

y_dest - The y position of the top-left corner of the destination image.

transparent - The alpha image.

src - The source rectangle.

fn putimage_alphatransparent( &mut self, image: &Image, x_dest: i32, y_dest: i32, transparent: impl IntoARGB, alpha: u8, src: Rect, ) -> Result<(), ImageError>

Draw an image with alpha blending and a specified alpha type.

Parameters

image - The image to draw.

x_dest - The x position of the top-left corner of the destination image.

y_dest - The y position of the top-left corner of the destination image.

transparent - The alpha value.

alpha - The alpha value.

src - The source rectangle.

fn putimage_rotate( &mut self, image: &Image, x_dest: i32, y_dest: i32, x_center: f32, y_center: f32, radian: f32, use_alpha: bool, alpha: Option<u8>, smooth: bool, ) -> Result<(), ImageError>

Draw an image with rotation.

Parameters

image - The image to draw.

x_dest - The x position of the top-left corner of the destination image.

y_dest - The y position of the top-left corner of the destination image.

x_center - The x position of the center of the rotation.

y_center - The y position of the center of the rotation.

radian - The rotation angle in radian.

use_alpha - Whether to use alpha blending.

alpha - The alpha value.

smooth - Whether to use smooth scaling.

fn putimage_rotatezoom( &mut self, image: &Image, x_dest: i32, y_dest: i32, x_center: f32, y_center: f32, radian: f32, zoom: f32, use_alpha: bool, alpha: Option<u8>, smooth: bool, ) -> Result<(), ImageError>

Draw an image with rotation and scale.

Parameters

image - The image to draw.

x_dest - The x position of the top-left corner of the destination image.

y_dest - The y position of the top-left corner of the destination image.

x_center - The x position of the center of the rotation.

y_center - The y position of the center of the rotation.

radian - The rotation angle in radian.

zoom - The zoom factor.

use_alpha - Whether to use alpha blending.

alpha - The alpha value.

smooth - Whether to use smooth scaling.

fn putimage_rotatetransparent( &mut self, image: &Image, x_center_dest: i32, y_center_dest: i32, src: Rect, x_center_src: i32, y_center_src: i32, transparent: impl IntoARGB, radian: f32, zoom: f32, ) -> Result<(), ImageError>

Draw an image with rotation and scale with alpha blending.

Parameters

image - The image to draw.

x_center_dest - The x position of the center of the destination image.

y_center_dest - The y position of the center of the destination image.

src - The source rectangle.

x_center_src - The x position of the center of the source image.

y_center_src - The y position of the center of the source image.

transparent - The alpha image.

radian - The rotation angle in radian.

zoom - The zoom factor.

fn imagefilter_blurring( &mut self, intensity: i32, alpha: u32, dest: Option<Rect>, ) -> Result<(), ImageError>

Blur the image.

Parameters

intensity - The intensity of the blur.

alpha - The light value.

dest - The destination rectangle.

fn drawimage(&mut self, image: &Image, x: i32, y: i32)

Draw an image with a specified position.

Parameters

image - The image to draw.

x - The x position.

y - The y position.

fn drawimage_with_scale(&mut self, image: &Image, dest: Rect, src: Rect)

Draw an image with scale.

Parameters

image - The image to draw.

dest - The destination rectangle.

src - The source rectangle.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.